Split the public API into three modules.
This simplifies the interface slightly as users can import only what
they need (e.g., if they only care about hashing in their script they
can just import the hashing API).

This is just the first stage, I actually want the names to be `hashing`,
`signing` and `verification`, but cannot do that at the moment, given we
already have `hashing` as a directory. I'll send another PR to rename
that to `_hashing` when we control more clearly what is the internal and
what is the external API and then will clean this up. But first, I need
to make sure the public API supports all of our needs for V1.
