+++ src/model_signing/hash.py
@@ -12,11 +12,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Public, high-level API for the model_signing library.
+"""High level API for the hashing interface of model_signing library.
 
-Users should use this API to sign models and verify the model integrity instead
-of reaching out to the internals of the library. We guarantee backwards
-compatibility only for the API defined in this file.
+Hashing is used both for signing and verification and users should ensure that
+the same configuration is used in both cases.
+
+Users should use this API to hash models (no signing and verification), rather
+than using the internals of the library. We guarantee backwards compatibility
+only for the API defined in `hash.py`, `sign.py` and `verify.py` at the root
+level of the library.
 """
 
 from collections.abc import Callable, Iterable
@@ -31,9 +35,6 @@ from model_signing.hashing import memory
 from model_signing.manifest import manifest
 from model_signing.serialization import serialize_by_file
 from model_signing.serialization import serialize_by_file_shard
-from model_signing.signing import in_toto
-from model_signing.signing import sign_sigstore as sigstore
-from model_signing.signing import signing
 
 
 if sys.version_info >= (3, 11):
@@ -61,43 +62,6 @@ def hash(model_path: os.PathLike) -> manifest.Manifest:
         A manifest of the hashed model.
     """
     return HashingConfig().hash(model_path)
-
-
-def sign(model_path: os.PathLike, signature_path: os.PathLike):
-    """Signs a model using the default configuration.
-
-    Args:
-        model_path: the path to the model to sign.
-        signature_path: the path of the resulting signature.
-    """
-    SigningConfig().sign(model_path, signature_path)
-
-
-def verify(
-    model_path: os.PathLike,
-    signature_path: os.PathLike,
-    *,
-    identity: str,
-    oidc_issuer: Optional[str] = None,
-    use_staging: bool = False,
-):
-    """Verifies that a model conforms to a signature.
-
-    Currently, this assumes signatures over DSSE, using Sigstore. We will add
-    support for more cases in a future change.
-
-    Args:
-        model_path: the path to the model to verify.
-        signature_path: the path to the signature to check.
-        identity: The expected identity that has signed the model.
-        oidc_issuer: The expected OpenID Connect issuer that provided the
-          certificate used for the signature.
-        use_staging: Use staging configurations, instead of production. This
-          is supposed to be set to True only when testing. Default is False.
-    """
-    VerificationConfig().set_sigstore_dsse_verifier(
-        identity=identity, oidc_issuer=oidc_issuer, use_staging=use_staging
-    ).verify(model_path, signature_path)
 
 
 class HashingConfig:
@@ -395,177 +359,4 @@ class HashingConfig:
             The new hashing configuration with a new set of ignored paths.
         """
         self._ignored_paths = frozenset({pathlib.Path(p) for p in paths})
-        return self
-
-
-class SigningConfig:
-    """Configuration to use when signing models.
-
-    The signing configuration is used to decouple between serialization formats
-    and signing types. This configuration class allows setting up the
-    serialization format, the method to convert a `manifest.Manifest` to a
-    signing payload and the engine used for signing (currently, only supporting
-    Sigstore at this level).
-    """
-
-    def __init__(self):
-        """Initializes the default configuration for signing."""
-        self._hashing_config = HashingConfig()
-        self._payload_generator = in_toto.DigestsIntotoPayload.from_manifest
-        self._signer = sigstore.SigstoreDSSESigner(
-            use_ambient_credentials=False, use_staging=False
-        )
-
-    def sign(self, model_path: os.PathLike, signature_path: os.PathLike):
-        """Signs a model using the current configuration.
-
-        Args:
-            model_path: the path to the model to sign.
-            signature_path: the path of the resulting signature.
-        """
-        manifest = self._hashing_config.hash(model_path)
-        payload = self._payload_generator(manifest)
-        signature = self._signer.sign(payload)
-        signature.write(pathlib.Path(signature_path))
-
-    def set_hashing_config(self, hashing_config: HashingConfig) -> Self:
-        """Sets the new configuration for hashing models.
-
-        Args:
-            hashing_config: the new hashing configuration.
-
-        Returns:
-            The new signing configuration.
-        """
-        self._hashing_config = hashing_config
-        return self
-
-    def set_payload_generator(
-        self, generator: Callable[[manifest.Manifest], signing.SigningPayload]
-    ) -> Self:
-        """Sets the conversion from manifest to signing payload.
-
-        Since we want to support multiple serialization formats and multiple
-        signing solutions, we use a payload generator to relax the coupling
-        between the two.
-
-        Args:
-            generator: the conversion function from a `manifest.Manifest` to a
-              `signing.SigningPayload` payload.
-
-        Return:
-            The new signing configuration.
-        """
-        self._payload_generator = generator
-        return self
-
-    def set_sigstore_signer(
-        self,
-        *,
-        oidc_issuer: Optional[str] = None,
-        use_ambient_credentials: bool = True,
-        use_staging: bool = False,
-        identity_token: Optional[str] = None,
-    ) -> Self:
-        """Configures the signing to be performed with Sigstore.
-
-        Only one signer can be configured. Currently, we only support Sigstore
-        in the API, but the CLI supports signing with PKI, BYOK and no signing.
-        We will merge the configurations in a subsequent change.
-
-        Args:
-            oidc_issuer: An optional OpenID Connect issuer to use instead of the
-              default production one. Only relevant if `use_staging = False`.
-              Default is empty, relying on the Sigstore configuration.
-            use_ambient_credentials: Use ambient credentials (also known as
-              Workload Identity). Default is True. If ambient credentials cannot
-              be used (not available, or option disabled), a flow to get signer
-              identity via OIDC will start.
-            use_staging: Use staging configurations, instead of production. This
-              is supposed to be set to True only when testing. Default is False.
-            identity_token: An explicit identity token to use when signing,
-              taking precedence over any ambient credential or OAuth workflow.
-
-        Return:
-            The new signing configuration.
-        """
-        self._signer = sigstore.SigstoreDSSESigner(
-            oidc_issuer=oidc_issuer,
-            use_ambient_credentials=use_ambient_credentials,
-            use_staging=use_staging,
-            identity_token=identity_token,
-        )
-        return self
-
-
-class VerificationConfig:
-    """Configuration to use when verifying models against signatures.
-
-    The signing configuration is used to decouple between serialization formats
-    and signing types. This configuration class allows setting up the
-    serialization format, the method to convert a `manifest.Manifest` to a
-    signing payload and the engine used for signing (currently, only supporting
-    Sigstore at this level).
-    """
-
-    def __init__(self):
-        """Initializes the default configuration for verification."""
-        self._hashing_config = HashingConfig()
-        self._verifier = None
-
-    def verify(self, model_path: os.PathLike, signature_path: os.PathLike):
-        """Verifies that a model conforms to a signature.
-
-        Args:
-            model_path: the path to the model to verify.
-            signature_path: the path to the signature to check.
-        """
-        signature = sigstore.SigstoreSignature.read(
-            pathlib.Path(signature_path)
-        )
-        expected_manifest = self._verifier.verify(signature)
-        actual_manifest = self._hashing_config.hash(model_path)
-
-        if actual_manifest != expected_manifest:
-            raise ValueError("Signature mismatch")
-
-    def set_hashing_config(self, hashing_config: HashingConfig) -> Self:
-        """Sets the new configuration for hashing models.
-
-        Args:
-            hashing_config: the new hashing configuration.
-
-        Returns:
-            The new signing configuration.
-        """
-        self._hashing_config = hashing_config
-        return self
-
-    def set_sigstore_dsse_verifier(
-        self,
-        *,
-        identity: str,
-        oidc_issuer: Optional[str] = None,
-        use_staging: bool = False,
-    ) -> Self:
-        """Configures the verification of a Sigstore signature over DSSE.
-
-        Only one verifier can be configured. Currently, we only support Sigstore
-        in the API, but the CLI supports signing with PKI, BYOK and no
-        signing/verification.  We will merge the configurations in a subsequent
-        change.
-
-        Args:
-            identity: The expected identity that has signed the model.
-            oidc_issuer: The expected OpenID Connect issuer that provided the
-              certificate used for the signature.
-            use_staging: Use staging configurations, instead of production. This
-              is supposed to be set to True only when testing. Default is False.
-
-        Return:
-            The new verification configuration.
-        """
-        self._verifier = sigstore.SigstoreDSSEVerifier(
-            identity=identity, oidc_issuer=oidc_issuer, use_staging=use_staging
-        )
         return self
